Fix concatenation bug when handling multi line entries in freedb record.
The original implementation was adding a space when concatenating
multiple lines (this was intended behaviour, tested in the unit tests).
My understanding, from reading the freedb record specs is that this is
the wrong behaviour and that the space should not be added.
(see EXDT line which is split on the word 'for' in the spec document
which can be found at http://ftp.freedb.org/pub/freedb/latest/DBFORMAT)