Fix for issue #608. Correctly rip hidden tracks.
Modify track selection to add hidden track (track 0) if ripping hidden
tracks is enabled and the hidden track exist (is detected).
The fix is not ideal because, the hidden track will be added even if
the user selects only certain tracks for ripping (not all of them).
To exclude the hidden track from ripping the user needs to disable
hidden track detection in preferences which is not really intuitive.